Polishing, renaming Maven and packages
