+++ .gitignore
@@ -1,2 +1,4 @@
 /core/target
 /test-app/target
+/test-app/*.iml
+/core/*.iml
