+++ .idea/modules.xml
@@ -2,6 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/core/fastcallhierarchy-core.iml" filepath="$PROJECT_DIR$/core/fastcallhierarchy-core.iml" />
+      <module fileurl="file://$PROJECT_DIR$/test-app/fastcallhierarchy-testproject.iml" filepath="$PROJECT_DIR$/test-app/fastcallhierarchy-testproject.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/java-fastest-callhierarchy.iml" filepath="$PROJECT_DIR$/.idea/java-fastest-callhierarchy.iml" />
       <module fileurl="file://$PROJECT_DIR$/core/superfastscanner.iml" filepath="$PROJECT_DIR$/core/superfastscanner.iml" />
       <module fileurl="file://$PROJECT_DIR$/test-app/testmodule.iml" filepath="$PROJECT_DIR$/test-app/testmodule.iml" />
