+++ core/pom.xml
@@ -2,13 +2,13 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>cz.george</groupId>
-		<artifactId>superfastscanner</artifactId>
+		<groupId>kacirekj</groupId>
+		<artifactId>fastcallhierarchy</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
 
-	<groupId>cz.george</groupId>
-	<artifactId>core</artifactId>
+	<groupId>kacirekj</groupId>
+	<artifactId>fastcallhierarchy-core</artifactId>
 	<version>1.0.1</version>
 	<packaging>jar</packaging>
 
@@ -23,23 +23,7 @@
 			<artifactId>asm-commons</artifactId>
 			<version>5.1</version>
 		</dependency>
-
-		<!-- Override dependency resolver with latest version.
-         This must go *BEFORE* the Arquillian BOM. -->
-		<dependency>
-			<groupId>org.jboss.shrinkwrap.resolver</groupId>
-			<artifactId>shrinkwrap-resolver-bom</artifactId>
-			<version>RELEASE</version>
-			<scope>import</scope>
-			<type>pom</type>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.12</version>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
 
-    
+
 </project>
