--- core/src/main/java/cz/george/superfastscanner/datastructures/ArrayListHashMap.java
@@ -1,11 +1,0 @@
-package cz.george.superfastscanner.datastructures;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ArrayListHashMap<Key, Value> extends OneToManyHashMap<Key, Value, List<Value>> {
-    @Override
-    protected List createCollectionInstance() {
-        return new ArrayList<>();
-    }
-}
