--- core/src/main/java/cz/george/superfastscanner/datastructures/HashMapHashSet.java
@@ -1,12 +1,0 @@
-package cz.george.superfastscanner.datastructures;
-import java.util.*;
-
-/**
- * HashMap<Key, HashSet<Value> >
- */
-public class HashMapHashSet<Key, Value> extends OneToManyHashMap<Key, Value, Set<Value>> {
-    @Override
-    protected Set createCollectionInstance() {
-        return new HashSet<>();
-    }
-}
