--- core/src/main/java/cz/george/superfastscanner/hierarchytree/package-info.java
@@ -1,3 +1,0 @@
-package cz.george.superfastscanner.hierarchytree;
-
-// This package should be refacored and removed.
