--- core/src/main/java/cz/george/superfastscanner/parsedbytecode/hashmap/ClassesHashMap.java
@@ -1,16 +1,0 @@
-package cz.george.superfastscanner.parsedbytecode.hashmap;
-
-import cz.george.superfastscanner.datastructures.OneToOneHashMap;
-import cz.george.superfastscanner.parsedbytecode.clazz.Clazz;
-
-import java.util.Set;
-
-/**
- * Class-oriented hashmap
- */
-public class ClassesHashMap extends OneToOneHashMap<Clazz, Clazz> {
-    public ClassesHashMap(Set<Clazz> classes) {
-        for(Clazz clazz : classes)
-            map.put(clazz, clazz);
-    }
-}
