+++ core/src/main/java/kacirekj/fastcallhierarchy/hierarchytree/package-info.java
@@ -1,0 +1,3 @@
+package kacirekj.fastcallhierarchy.hierarchytree;
+
+// This package should be refacored and removed.
