+++ pom.xml
@@ -3,9 +3,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
                              http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-		
-	<groupId>cz.george</groupId>
-	<artifactId>superfastscanner</artifactId>
+
+	<groupId>kacirekj</groupId>
+	<artifactId>fastcallhierarchy</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 
 	<packaging>pom</packaging>
@@ -36,7 +36,6 @@
 		</plugins>
 	</build>
 
-
 	<dependencies>
         <dependency>
             <groupId>org.projectlombok</groupId>
@@ -53,7 +52,7 @@
 	</dependencies>
 
 	<modules>
-        <module>core</module>
-        <module>test-app</module>
+		<module>core</module>
+		<module>test-app</module>
     </modules>
 </project>
